requirements.txt
numpy>=1.25.0
scipy>=1.11.0
pyyaml>=6.0
matplotlib>=3.7.0
pytest>=7.4.0
